let's try with another knn method
git-svn-id: http://google-refine.googlecode.com/svn/trunk@254 7d457c2a-affb-35e4-300a-418c747d4874

+++ src/main/java/edu/mit/simile/vicino/distances/BZip2Distance.java
@@ -9,7 +9,7 @@ public class BZip2Distance extends PseudoMetricDistance {
 
     public double d2(String x, String y) {
         String str = x + y;
-        float result = 0.0f;
+        double result = 0.0f;
         try {
             ByteArrayOutputStream baos = new ByteArrayOutputStream(str.length());
             CBZip2OutputStream os = new CBZip2OutputStream(baos);
