+++ src/main/java/edu/mit/simile/vicino/distances/GZipDistance.java
@@ -8,7 +8,7 @@ public class GZipDistance extends PseudoMetricDistance {
 
     public double d2(String x, String y) {
         String str = x + y;
-        float result = 0.0f;
+        double result = 0.0f;
         try {
             ByteArrayOutputStream baos = new ByteArrayOutputStream(str.length());
             GZIPOutputStream os = new GZIPOutputStream(baos);
